tests: expand Open tests to include various token pairs'
