+++ packages/oasis-actions/src/strategies/types/IOperation.ts
@@ -1,0 +1,7 @@
+import { ActionCall } from '../../actions/types/actionCall'
+import { OperationNames } from '../../helpers/constants'
+
+export interface IOperation {
+  calls: ActionCall[]
+  operationName: OperationNames
+}
